+++ tests/test_conn_interface.c
@@ -102,8 +102,9 @@ void set_global_shutdown_reason(char *reason)
     UNUSED(reason);
 }
 
-void start_conn_in_progress (void)
+void start_conn_in_progress (unsigned long start_time)
 {
+	UNUSED(start_time);
 }   
 
 void stop_conn_in_progress (void)