+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - send status code and reason in websocket close message
 - dont try to install handler for signal 9
 - on connect retry, requery jwt only if it failed before 
+- put two timestamps in connection health file; start conn and current
 
 ## [1.0.2] - 2019-02-08
 - Refactored connection.c and updated corresponding unit tests